requirements.txt
flask>=3.1,<4
requests>=2.32,<3
paho-mqtt>=1.6,<2
